spin off image upload creation into an assignments api
