+++ tests/feature/api/test_search.py
@@ -1,14 +1,8 @@
 from unittest.mock import call
-from io import BytesIO
-from PIL import Image
-import filecmp
 import random
 
 from db.models import Lecture
 import api.routers.search
-from db.crud import (
-    get_image_upload_by_public_id,
-)
 
 
 def test_api_can_search_courses(mocker, api_client):
@@ -243,97 +237,6 @@ def test_lecture_transcript_search(mocker, api_client, analysed_lecture):
     assert len(response.json()) == 1
     assert index.call_count == 1
     assert response.json()[0]['highlight']['transcript'][0] == '00:00 -> 00:30 foo'
-
-
-def test_image_search_creates_image_upload(mocker, api_client, img_file):
-    mocker.patch('jobs.schedule_image_search')
-
-    response = api_client.post(
-        '/search/image',
-        files={
-            'file': ('a_file_name.png', open(img_file, 'rb'), 'image/png')
-        }
-    )
-
-    image_id = response.json()['id']
-
-    upload = get_image_upload_by_public_id(image_id)
-
-    assert upload is not None
-
-
-def test_image_search_will_not_save_the_same_image_twice(mocker, api_client, img_file):
-    mocker.patch('jobs.schedule_image_search')
-
-    def func():
-        response = api_client.post(
-            '/search/image',
-            files={
-                'file': ('a_file_name.png', open(img_file, 'rb'), 'image/png')
-            }
-        )
-        return response
-
-    response = func()
-    image_id_1 = response.json()['id']
-
-    response = func()
-    image_id_2 = response.json()['id']
-
-    assert image_id_1 == image_id_2
-
-
-def test_image_search_saves_png_file_upload(mocker, api_client, img_file):
-    mocker.patch('jobs.schedule_image_search')
-
-    response = api_client.post(
-        '/search/image',
-        files={
-            'file': ('a_file_name.png', open(img_file, 'rb'), 'image/png')
-        }
-    )
-
-    image_id = response.json()['id']
-
-    upload = get_image_upload_by_public_id(image_id)
-    assert filecmp.cmp(img_file, upload.get_filename())
-
-
-def test_image_search_can_retrieve_image(api_client, image_upload):
-    response = api_client.get(f'/search/image/{image_upload.public_id}/img')
-
-    image = Image.open(BytesIO(response.content))
-    assert image.format == 'PNG'
-
-
-def test_image_search_can_retrieve_info(api_client, image_upload):
-    image_upload.text_content = 'foo'
-    image_upload.description_en = 'bar'
-    image_upload.description_sv = 'baz'
-    image_upload.save()
-
-    response = api_client.get(f'/search/image/{image_upload.public_id}')
-
-    assert response.json()['text_content'] == 'foo'
-    assert response.json()['description_en'] == 'bar'
-    assert response.json()['description_sv'] == 'baz'
-
-
-def test_image_upload_schedules_image_search_pipeline(mocker, api_client, img_file):
-    schedule_image_search_mock = mocker.patch('jobs.schedule_image_search')
-
-    response = api_client.post(
-        '/search/image',
-        files={
-            'file': ('a_file_name.png', open(img_file, 'rb'), 'image/png')
-        }
-    )
-
-    image_id = response.json()['id']
-    image = get_image_upload_by_public_id(image_id)
-
-    assert schedule_image_search_mock.call_count == 1
-    assert schedule_image_search_mock.mock_calls[0] == call(image)
 
 
 def test_image_question_can_be_created_for_uploaded_image(mocker, api_client, image_upload):
